spike(sealed-guest): split guest into its own crate — 12.7MB → 2.3MB
The guest bin lived in tonk-ui, so it inherited tonk-ui's dependency on
tonk-worker — dragging the entire service worker (analyzer, evaluator,
language-server, reactor, repository, storage) into the sealed-guest wasm.
That stuff belongs across the iframe boundary, not in the view runtime.

Move the guest into a standalone tonk-guest crate that depends only on the
custom-element crates (tonk-display/board/workspace/tree/sigil) + the proxy
host. The guest wasm drops from 12.7MB to 2.3MB (release+wasm-opt -Oz: 2.5MB),
and the tonk-worker snippet import disappears. Re-verified in the sealed
opaque-origin iframe: tonk-display / tonk-host / tonk-board all register,
origin "null", upgraded.

The remaining ~2.3MB is the element crates' dialog-artifacts→dialog-storage
chain (Conclusion/Query wire types anchoring the DB types crate); slimming
that further is a separate, optional refactor.
