+++ rust/tonk-guest/src/guest_host.rs
